pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "email-triage-env"
version = "0.1.0"
description = "FastAPI email triage environment with OpenEnv-compatible submission runner."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi",
  "uvicorn",
  "pydantic",
  "openai",
  "openenv-core",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
py-modules = ["app", "inference"]

[tool.setuptools.packages.find]
include = ["env*", "server*"]